Refactor: moduleDomain에서 파일 위치 변경에 따른 경로 수정
