+++ module-api/src/main/java/com/example/moduleapi/service/auction/AuctionService.java
@@ -9,11 +9,11 @@ import com.example.moduleapi.exception.auction.RedisLockInterruptedException;
 import com.example.moduleapi.service.dto.BiddingSuccessfulResult;
 import com.example.moduleapi.service.point.PointService;
 import com.example.moduleapi.service.product.ProductFacade;
+import com.example.moduledomain.common.response.ProductFindResponse;
 import com.example.moduledomain.domain.bidLogging.BidLogging;
 import com.example.moduledomain.domain.user.CustomUserDetails;
 import com.example.moduledomain.domain.user.Gender;
 import com.example.moduledomain.domain.user.User;
-import com.example.moduledomain.response.ProductFindResponse;
 import org.redisson.api.RLock;
 import org.redisson.api.RMap;
 import org.redisson.api.RedissonClient;
