+++ module-recommendation/src/main/java/com/example/modulerecommendation/controller/ProductRecommendationController.java
@@ -1,17 +1,12 @@
 package com.example.modulerecommendation.controller;
 
-import java.util.List;
-
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
+import com.example.moduledomain.common.request.ProductFilter;
+import com.example.moduledomain.common.response.ProductFindResponse;
 import com.example.moduledomain.domain.user.Gender;
-import com.example.moduledomain.request.ProductFilter;
-import com.example.moduledomain.response.ProductFindResponse;
 import com.example.modulerecommendation.service.ProductRecommendationService;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/products")
@@ -24,9 +19,9 @@ public class ProductRecommendationController {
 
     @PostMapping("/_recommendations")
     public List<ProductFindResponse> findRecommendationProducts(
-        @RequestParam Gender gender,
-        @RequestParam int age,
-        @RequestBody ProductFilter productFilter) {
+            @RequestParam Gender gender,
+            @RequestParam int age,
+            @RequestBody ProductFilter productFilter) {
         return productRecommendationService.getRecommendationProducts(gender, age, productFilter);
     }
 }
